refactor: ディレクトリ配置を最適化しました。
