+++ app/src/main/java/jp/panta/misskeyandroidclient/di/module/GalleryModule.kt
@@ -5,8 +5,8 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import net.pantasystem.milktea.data.infrastructure.gallery.GalleryPostsStoreImpl
-import net.pantasystem.milktea.data.infrastructure.gallery.impl.GalleryRepositoryImpl
-import net.pantasystem.milktea.data.infrastructure.gallery.impl.InMemoryGalleryDataSource
+import net.pantasystem.milktea.data.infrastructure.gallery.GalleryRepositoryImpl
+import net.pantasystem.milktea.data.infrastructure.gallery.InMemoryGalleryDataSource
 import net.pantasystem.milktea.model.gallery.GalleryDataSource
 import net.pantasystem.milktea.model.gallery.GalleryPostsStore
 import net.pantasystem.milktea.model.gallery.GalleryRepository
