+++ app/src/main/java/jp/panta/misskeyandroidclient/di/module/UserModule.kt
@@ -4,9 +4,9 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import net.pantasystem.milktea.data.infrastructure.user.impl.FollowFollowerPagingStoreImpl
-import net.pantasystem.milktea.data.infrastructure.user.impl.InMemoryUserDataSource
-import net.pantasystem.milktea.data.infrastructure.user.impl.UserRepositoryImpl
+import net.pantasystem.milktea.data.infrastructure.user.FollowFollowerPagingStoreImpl
+import net.pantasystem.milktea.data.infrastructure.user.InMemoryUserDataSource
+import net.pantasystem.milktea.data.infrastructure.user.UserRepositoryImpl
 import net.pantasystem.milktea.model.user.FollowFollowerPagingStore
 import net.pantasystem.milktea.model.user.UserDataSource
 import net.pantasystem.milktea.model.user.UserRepository
