+++ app/src/main/java/jp/panta/misskeyandroidclient/di/module/UserNicknameModule.kt
@@ -4,9 +4,9 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import net.pantasystem.milktea.data.infrastructure.user.impl.UserNicknameDAO
-import net.pantasystem.milktea.data.infrastructure.user.impl.UserNicknameRepositoryOnMemoryImpl
-import net.pantasystem.milktea.data.infrastructure.user.impl.UserNicknameRepositorySQLiteImpl
+import net.pantasystem.milktea.data.infrastructure.user.UserNicknameDAO
+import net.pantasystem.milktea.data.infrastructure.user.UserNicknameRepositoryOnMemoryImpl
+import net.pantasystem.milktea.data.infrastructure.user.UserNicknameRepositorySQLiteImpl
 import net.pantasystem.milktea.model.user.nickname.UserNicknameRepository
 import javax.inject.Singleton
