+++ app/src/test/java/jp/panta/misskeyandroidclient/model/users/nickname/DeleteNicknameUseCaseTest.kt
@@ -5,8 +5,8 @@ import jp.panta.misskeyandroidclient.logger.TestLogger
 import jp.panta.misskeyandroidclient.model.account.TestAccountRepository
 
 import kotlinx.coroutines.runBlocking
-import net.pantasystem.milktea.data.infrastructure.user.impl.InMemoryUserDataSource
-import net.pantasystem.milktea.data.infrastructure.user.impl.UserNicknameRepositoryOnMemoryImpl
+import net.pantasystem.milktea.data.infrastructure.user.InMemoryUserDataSource
+import net.pantasystem.milktea.data.infrastructure.user.UserNicknameRepositoryOnMemoryImpl
 import net.pantasystem.milktea.model.account.Account
 import net.pantasystem.milktea.model.account.AccountRepository
 import net.pantasystem.milktea.model.user.User
