+++ data/src/main/java/net/pantasystem/milktea/data/gettters/MessageRelationGetterImpl.kt
@@ -2,7 +2,7 @@ package net.pantasystem.milktea.data.gettters
 
 import net.pantasystem.milktea.api.misskey.messaging.MessageDTO
 import net.pantasystem.milktea.data.infrastructure.entities
-import net.pantasystem.milktea.data.infrastructure.messaging.impl.MessageDataSource
+import net.pantasystem.milktea.data.infrastructure.messaging.MessageDataSource
 import net.pantasystem.milktea.data.infrastructure.toGroup
 import net.pantasystem.milktea.model.account.Account
 import net.pantasystem.milktea.model.account.AccountRepository
