+++ data/src/main/java/net/pantasystem/milktea/data/infrastructure/group/GroupRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package net.pantasystem.milktea.data.infrastructure.group.impl
+package net.pantasystem.milktea.data.infrastructure.group
 
 
 import net.pantasystem.milktea.api.misskey.groups.*
