+++ data/src/main/java/net/pantasystem/milktea/data/infrastructure/messaging/MessageDataSource.kt
@@ -1,4 +1,4 @@
-package net.pantasystem.milktea.data.infrastructure.messaging.impl
+package net.pantasystem.milktea.data.infrastructure.messaging
 
 
 import kotlinx.coroutines.flow.Flow
