+++ data/src/main/java/net/pantasystem/milktea/data/infrastructure/messaging/MessageObserverImpl.kt
@@ -1,4 +1,4 @@
-package net.pantasystem.milktea.data.infrastructure.messaging.impl
+package net.pantasystem.milktea.data.infrastructure.messaging
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
