+++ data/src/main/java/net/pantasystem/milktea/data/infrastructure/streaming/StreamingMainMessageEventDispatcher.kt
@@ -2,7 +2,7 @@ package net.pantasystem.milktea.data.infrastructure.streaming
 
 import net.pantasystem.milktea.data.gettters.MessageAdder
 import net.pantasystem.milktea.model.messaging.MessageRelationGetter
-import net.pantasystem.milktea.data.infrastructure.messaging.impl.MessageDataSource
+import net.pantasystem.milktea.data.infrastructure.messaging.MessageDataSource
 import net.pantasystem.milktea.data.streaming.ChannelBody
 import net.pantasystem.milktea.model.account.Account
 import javax.inject.Inject
