+++ data/src/main/java/net/pantasystem/milktea/data/infrastructure/user/UserNicknameRepositorySQLiteImpl.kt
@@ -1,4 +1,4 @@
-package net.pantasystem.milktea.data.infrastructure.user.impl
+package net.pantasystem.milktea.data.infrastructure.user
 
 import net.pantasystem.milktea.model.user.nickname.UserNickname
 import net.pantasystem.milktea.model.user.nickname.UserNicknameNotFoundException
